+++ grails-app/domain/org/pih/warehouse/inventory/CycleCountDetails.groovy
@@ -25,18 +25,18 @@ class CycleCountDetails implements Serializable {
     // Blind count details
     Date blindCountDateCounted
     Person blindCountAssignee
-    Integer blindCountQuantityOnHand
-    Integer blindCountQuantityCounted
-    Integer blindCountQuantityVariance
+    BigDecimal blindCountQuantityOnHand
+    BigDecimal blindCountQuantityCounted
+    BigDecimal blindCountQuantityVariance
     String blindCountVarianceReasonCode
     String blindCountVarianceComment
 
     // Verification count details
     Date verificationCountDateCounted
     Person verificationCountAssignee
-    Integer verificationCountQuantityOnHand
-    Integer verificationCountQuantityCounted
-    Integer verificationCountQuantityVariance
+    BigDecimal verificationCountQuantityOnHand
+    BigDecimal verificationCountQuantityCounted
+    BigDecimal verificationCountQuantityVariance
     String verificationCountVarianceReasonCode
     String verificationCountVarianceComment
 
@@ -51,7 +51,7 @@ class CycleCountDetails implements Serializable {
         table "cycle_count_details"
     }
 
-    VarianceTypeCode getVarianceTypeCode(Integer quantityVariance) {
+    VarianceTypeCode getVarianceTypeCode(BigDecimal quantityVariance) {
         if (quantityVariance > 0) return VarianceTypeCode.MORE
         else if (quantityVariance < 0) return VarianceTypeCode.LESS
         else return VarianceTypeCode.EQUAL
